src/backend/summarization/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic[email]==2.6.3
requests==2.31.0

# Mongo
pymongo==4.6.0

# Celery + RabbitMQ
celery==5.3.6

# Transformers Summarization
transformers==4.38.2
torch==2.2.1

# Dotenv
python-dotenv==1.0.1

# Testing
pytest==8.0.0
httpx==0.26.0
pytest-cov==4.1.0

gunicorn==21.2.0
